+++ cmd/hidden_lake/applications/messenger/pkg/app/app.go
@@ -33,9 +33,9 @@ var (
 type sApp struct {
 	fState state.IState
 
-	fConfig config.IConfig
-	fPasw   string
-	fPathTo string
+	fConfig   config.IConfig
+	fPassword string
+	fPathTo   string
 
 	fDatabase       database.IKVDatabase
 	fIntServiceHTTP *http.Server
@@ -48,7 +48,7 @@ type sApp struct {
 
 func NewApp(
 	pCfg config.IConfig,
-	pPasw string,
+	pPassword string,
 	pPathTo string,
 ) types.IRunner {
 	httpLogger := std_logger.NewStdLogger(pCfg.GetLogging(), http_logger.GetLogFunc())
@@ -57,7 +57,7 @@ func NewApp(
 	return &sApp{
 		fState:      state.NewBoolState(),
 		fConfig:     pCfg,
-		fPasw:       pPasw,
+		fPassword:   pPassword,
 		fPathTo:     pPathTo,
 		fHTTPLogger: httpLogger,
 		fStdfLogger: stdfLogger,