+++ cmd/hidden_lake/applications/messenger/pkg/app/init_database.go
@@ -12,7 +12,7 @@ import (
 func (p *sApp) initDatabase() error {
 	sett := database.NewSettings(&database.SSettings{
 		FPath:     filepath.Join(p.fPathTo, hlm_settings.CPathDB),
-		FPassword: p.fPasw,
+		FPassword: p.fPassword,
 		FWorkSize: p.fConfig.GetSettings().GetWorkSizeBits(),
 	})
 	db, err := hlm_database.NewKeyValueDB(sett)