+++ cmd/hidden_lake/composite/pkg/app/init_app.go
@@ -39,7 +39,6 @@ func InitApp(
 	pArgs []string,
 	pDefaultPath string,
 	pDefaultKey string,
-	pDefaultPasw string,
 	pDefaultParallel uint64,
 ) (types.IRunner, error) {
 	inputPath := strings.TrimSuffix(flag.GetFlagValue(pArgs, "path", pDefaultPath), "/")
@@ -49,7 +48,7 @@ func InitApp(
 		return nil, utils.MergeErrors(ErrInitConfig, err)
 	}
 
-	runners, err := getRunners(cfg, pArgs, pDefaultPath, pDefaultKey, pDefaultPasw, pDefaultParallel)
+	runners, err := getRunners(cfg, pArgs, pDefaultPath, pDefaultKey, pDefaultParallel)
 	if err != nil {
 		return nil, utils.MergeErrors(ErrGetRunners, err)
 	}
@@ -62,7 +61,6 @@ func getRunners(
 	pArgs []string,
 	pDefaultPath string,
 	pDefaultKey string,
-	pDefaultPasw string,
 	pDefaultParallel uint64,
 ) ([]types.IRunner, error) {
 	var (
@@ -92,7 +90,7 @@ func getRunners(
 		case hll_settings.CServiceFullName:
 			runner, err = hll_app.InitApp(pArgs, pDefaultPath)
 		case hlm_settings.CServiceFullName:
-			runner, err = hlm_app.InitApp(pArgs, pDefaultPath, pDefaultPasw)
+			runner, err = hlm_app.InitApp(pArgs, pDefaultPath, pDefaultKey)
 		case hlf_settings.CServiceFullName:
 			runner, err = hlf_app.InitApp(pArgs, pDefaultPath)
 		case hla_common_settings.CServiceFullName: